[REFACTOR]: 폴더 구조 분리
