+++ src/main/java/com/example/ai_tutor/domain/practice/application/PracticeService.java
@@ -4,7 +4,11 @@ import com.example.ai_tutor.domain.note.domain.Note;
 import com.example.ai_tutor.domain.note.domain.repository.NoteRepository;
 import com.example.ai_tutor.domain.practice.domain.Practice;
 import com.example.ai_tutor.domain.practice.domain.repository.PracticeRepository;
-import com.example.ai_tutor.domain.practice.dto.*;
+import com.example.ai_tutor.domain.practice.dto.request.AnswerReq;
+import com.example.ai_tutor.domain.practice.dto.request.UpdateAnswersReq;
+import com.example.ai_tutor.domain.practice.dto.response.PracticeRes;
+import com.example.ai_tutor.domain.practice.dto.response.PracticeResultsRes;
+import com.example.ai_tutor.domain.practice.dto.response.TutorRecordRes;
 import com.example.ai_tutor.global.DefaultAssert;
 import com.example.ai_tutor.global.config.security.token.UserPrincipal;
 import com.example.ai_tutor.global.payload.ApiResponse;
