+++ src/main/java/com/example/ai_tutor/domain/practice/dto/response/PracticeRes.java
@@ -1,4 +1,4 @@
-package com.example.ai_tutor.domain.practice.dto;
+package com.example.ai_tutor.domain.practice.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
